Optimizations

+++ src/Chunk/ChunkMesh.hpp
@@ -21,17 +21,17 @@ namespace CrossCraft {
 
         ChunkMeshCollection mesh;
 
-        void add_block_to_mesh(const WorldData *wd, block_t block, const mathfu::Vector<float, 3> &vector3,
+        void add_block_to_mesh(const WorldData *wd, block_t block, const mathfu::Vector<int, 3> &vector3,
                                const SurroundingPositions &positions);
 
         void try_add_face(const WorldData *wd, const std::array<float, 12> &data, block_t block,
-                          const mathfu::Vector<float, 3> &actual_pos, const mathfu::Vector<int, 3> &check_pos,
-                          uint32_t lightValue);
+                          const mathfu::Vector<int, 3> &actual_pos, const mathfu::Vector<int, 3> &check_pos,
+                          uint16_t lightValue);
 
         static void
         add_face_to_mesh(const std::array<float, 12> &array1, std::array<float, 8> array2,
-                         const mathfu::Vector<float, 3> &vector3,
-                         uint32_t value,
+                         const mathfu::Vector<int, 3> &vector3,
+                         uint16_t value,
                          ChunkMeshInstance &instance);
     };
 }
